fix: remove styles folder
